+++ History.md
@@ -1,6 +1,10 @@
 <!-- HISTORY -->
 ## History
 
+* 0.37.15 - 21.12.2025
+  * Python upgrader on Windows fix.
+  * Added Rufus Portable installer (Windows).
+  * Remove empty folders in config dir.
 * 0.37.14 - 21.12.2025
   * Fixed Python installer and upgrader on Windows.
 * 0.37.13 - 21.12.2025